API: HTTP upstreams statistics.
The statistics is collected for upstreams with the "zone" directive.

This adds "/status/http/upstreams/<upstream>" objects with the following
fields:

 - "peers" (object): upstream peers statistics

 - "keepalive" (number): the number of currently cached connections

where <upstream> is the name of any upstream specified with the "zone"
directive.

The "peers" object contains the following fields for each peer:

 - <address> (object): peer's statistics

where <address> is the canonical text representation of the peer address.

The <address> object contains the following fields:

 - "server" (string): the parameter of the "server" directive

 - "backup" (boolean): true for backup servers

 - "weight" (number): configured weight

 - "state" (string): the current state of the server, can either be "up",
                     "down", or "unavailable" (when "max_fails" limit is
                     reached)

 - "selected" (object): peer selection statistics

 - "max_conns" (number): the configured maximum number of simultaneous
                         connections, if specified

 - "responses" (object): responses statistics

 - "data" (object): data statistics

 - "health" (object): health statistics

The "selected" object contains the following fields:

 - "current" (number): the current number of connections to peer

 - "total" (number): total number of requests forwarded to peer

 - "last" (string): time in ISO 8601 format when peer was last selected

The "responses" object contains the following fields:

 - "<code>" (number): a non-zero number of responses with status <code> (100-599)

 - "xxx" (number): a non-zero number of responses with other status codes

The "data" object contains the following fields:

 - "received" (number): the total number of bytes received from peer

 - "sent" (number): the total number of bytes sent to peer

The "health" object contains the following fields:

 - "fails" (number): the total number of unsuccessful attempts to communicate
                     with the peer

 - "unavailable" (number): how many times peer became unavailable due to
                           reaching the "max_fails" limit

 - "downtime" (number): the total time (in milliseconds) when peer was
                        unavailable for selection due to being "unavailable"

 - "downstart" (string): time in ISO 8601 format when peer became "unavailable"

The new macro NGX_HTTP_UPSTREAM_CONF explicitly marks balancers based
on round-robin (that collects upstreams statistics), and serves as the
prerequisite for the following changes.

The logic in ngx_http_upstream_free_round_robin_peer() was changed to
not differentiate between single and multiple peers cases for accounting
of "unavailable" state.