Tests: workaround for QUIC routing while reloading nginx.
Datagrams can be routed to a new worker process that lacks the connection
state resulting in a stateless reset.  Previously, the QUIC connection was
closed in tests on decryption failure including receiving a stateless reset,
which is not yet detected in HTTP3.pm.  The workaround is to ignore packets
that could not be decrypted including a stateless reset.  This is enough to
continue receiving packets from the old worker process, such as with GOAWAY.

--HG--
branch : nginx-tests