HTTP: implementation of the sticky sessions in upstream.
The module allows to "stick" client to a particular upstream server.
In the "cookie" mode this is achieved by setting cookie which value
identifies upstream server.

In the "route" mode, upstream server assigns client a route, and the
client carries this information in all further requests.  The server
may use this information to route the client to a particular server
identified by the "sid" parameter.

Example configuration:

upstream u_cookie {
    server 127.0.0.1:8081;
    server 127.0.0.1:8082;

    sticky cookie example domain=example.com max-age=3600;
}

upstream u_route {

    server 127.0.0.1:8081 sid=backend1;
    server 127.0.0.1:8082 sid=backend2;

    sticky route $http_route $cookie_route;

    sticky_strict on;
}