Tests: reworked stream SSL tests to use IO::Socket::SSL.
Relevant infrastructure is provided in Test::Nginx::Stream.  This also
ensures that SSL handshake and various read operations are guarded
with timeouts.

The stream_ssl_verify_client.t test uses IO::Socket::SSL::_get_ssl_object()
to access the Net::SSLeay object directly, as it seems to be the only
way to obtain CA list with IO::Socket::SSL.  While not exactly correct,
this seems to be good enough for tests.

--HG--
branch : nginx-tests