Tests: unbreak reading new stderr data after eof.
Tests don't expect to stop reading redirected stderr when end of file is
reached, but rather to read new data being appended, similar to "tail -f".
The behaviour is found changed in Ubuntu 23.04's Perl 5.36, which applies
the upstream patch [1] expected for inclusion in the upcoming Perl 5.38.
The fix is to clear the filehandle's error state to continue reading.

[1] https://github.com/Perl/perl5/commit/80c1f1e45e8e

Updated mail_error_log.t and stream_error_log.t for consistency.

--HG--
branch : nginx-tests