Upstream zone: reduced shared memory usage for peer with API.
Previously, the array of status code counters was part of the peer structure,
which increased its size slightly to over 4096 bytes.  As a result, each peer
consumed two slab pages, with the second page having a long tail of unused
memory.

The size of that array alone is 4008 bytes, and now it's allocated by a
separate slab memory request, consuming only one page.  All the other
allocations for peers are significantly smaller and can fit into individual
slab slots, which reduces the amount of wasted shared zone memory.

Additionally, the response code counters are not allocated and counted without
a configured shared memory zone.  The 'bytes sent/received' counters are also
not counted in this case.