API: stream upstreams statistics.
Statistics are collected for upstreams using the "zone" directive.

This adds "/status/stream/upstreams/<upstream>" objects with the following
fields:

 - "peers" (object): upstream peers statistics

where <upstream> is the name of any upstream specified with the "zone"
directive.

The "peers" object contains the following fields for each peer:

 - <address> (object): peer statistics

where <address> is the canonical text representation of the peer's address.

The <address> object contains the following fields:

 - "server" (string): the parameter of the "server" directive

 - "backup" (boolean): true for backup servers

 - "weight" (number): the configured weight

 - "state" (string): the current state of the server, can be either "up",
                     "down", or "unavailable" (if the "max_fails" limit
                     is reached)

 - "selected" (object): peer selection statistics

 - "max_conns" (number): the configured maximum number of concurrent
                         connections, if specified

 - "data" (object): data statistics

 - "health" (object): health statistics

The "selected" object contains the following fields:

 - "current" (number): the current number of connections to the peer

 - "total" (number): total number of connections forwarded to peer

 - "last" (string): time in ISO 8601 format when peer was last selected

The "data" object contains the following fields:

 - "received" (number): the total number of bytes received from peer

 - "sent" (number): the total number of bytes sent to peer

The "health" object contains the following fields:

 - "fails" (number): the total number of unsuccessful attempts to communicate
                     with the peer

 - "unavailable" (number): the number of times the peer has been unavailable due
                           to reaching the "max_fails" limit

 - "downtime" (number): the total time (in milliseconds) that the peer wasn't
                        available for selection due to being "unavailable"

 - "downstart" (string): time in ISO 8601 format when peer became "unavailable"

The new macro NGX_STREAM_UPSTREAM_CONF explicitly flags balancers using the
round-robin method (which collects upstream statistics), and serves as a
prerequisite for the following changes.

The logic in ngx_stream_upstream_free_round_robin_peer() was changed to
not differentiate between single and multiple peer cases when accounting
for "unavailable" states.