Prometheus exporter module.
This introduces two directives:

 - "prometheus <NAME>" in the location context, which sets the content handler
   in a surrounding location and outputs metrics in Prometheus format according
   to the template specified by name in the first mandatory parameter

 - "prometheus_template <NAME> { ... }" in the context of the http block that
   specifies the template block that configures all metrics produced using this
   template name in the "prometheus" directive

The template block can contain multiple metric directives with the following
format:

  <NAME> $<VARIABLE_VALUE> [path=<MATCHING_PATH>] [type=<TYPE>] [help=<HELP>]

The <NAME> and <VARIABLE_VALUE> positional parameters are mandatory, all others
are optional.

<NAME> specifies the Prometheus metric name with optional labels part, e.g.:

  'angie_slabs_slots_free{zone="$1",size="$2"}'

The labels part in {} can contain variables, including those defined by the
<MATCHING_PATH> with regexp.

<VARIABLE_VALUE> specifies the variable that will be evaluated to get the
current metric value.  The evaluation is done at the same time as the evaluation
of the variables in the labels part of the metric name.  If the variable is
evaluated to an empty value or is not found, then the corresponding metric
isn't output.

The optional <MATCHING_PATH> specifies a path that will be matched against all
existing full metric paths in the /status/ API subtree (with a leading slash
and no trailing slash):

  /angie/version
  /angie/address
  /angie/generation
  /angie/load_time
  /connections/accepted
  /connections/dropped
  /connections/active
  /connections/idle
  /slabs/one/pages/used

...and so on.

It's only available when the API module is built.

There are two matching methods: exact match and regex.  In the latter case, the
<MATCHING_PATH> must start with ~ and is only available with PCRE.  Examples
include:

  path=/angie/generation
  path=~^/http/upstreams/([^/]+)/peers/([^/]+)/state$

The regex may contain captures that can be used in the labels part of <NAME> or
during <VARIABLE_VALUE> evaluation.

For each matching API subtree path, the <NAME> metric is returned and the
special variable $p8s_value is set to the corresponding metric value.

This allows you to print metrics with the same name but different labels, for
example:

  'angie_slabs_slots_free{zone="$1",size="$2"}' $p8s_value
        path=~^/slabs/([^/]+)/slots/([^/]+)/free$;

Returns all free slot page counts for different zones and slab sizes:

  angie_slabs_slots_free{zone="one",size="8"} 502
  angie_slabs_slots_free{zone="one",size="16"} 249
  angie_slabs_slots_free{zone="one",size="32"} 122
  angie_slabs_slots_free{zone="one",size="128"} 22
  angie_slabs_slots_free{zone="one",size="512"} 4

If there are no matches, then the corresponding metric isn't returned at all.

The $p8s_value can be converted to a Prometheus-meaningful metric value using
scripting or the map directive.

For example, this is useful for converting textual upstream states to numbers:

    map $p8s_value $ups_state_n {
        up           0;
        unavailable  1;
        down         2;
        default      3;
    }

    prometheus_template main {
       'angie_http_upstreams_state{upstream="$1",peer="$2"}' $ups_state_n
            path=~^/http/upstreams/([^/]+)/peers/([^/]+)/state$;
    }

For the boolean type values, $p8s_value is set to "0" or "1" respectively.
For nulls, it is set to "(null)".

If no <MATCHING_PATH> is specified, then the <NAME> metric is returned just
once (with non-empty evaluation of <VARIABLE_VALUE>).

The optional <TYPE> and <HELP> parameters simply set the Prometheus metric
"TYPE" and metric "HELP" strings for a group of metrics with the corresponding
metric <NAME> (without labels).  For example (using the map directive above):

   'peer_state{upstream="$1",peer="$2"}' $ups_state_n
        path=~^/http/upstreams/([^/]+)/peers/([^/]+)/state$
        type=gauge
        "help=upstream peer state 0 - up, 1 - unavail, 2 - down, 3 - other";

    current_time $msec type=counter "help=current time in seconds";

Will produce output similar to the following:

  # HELP peer_state upstream peer state 0 - up, 1 - unavail, 2 - down, 3 - other
  # TYPE peer_state gauge
  peer_state{upstream="one",peer="127.0.0.1:80"} 0
  peer_state{upstream="one",peer="127.0.0.2:80"} 2
  peer_state{upstream="one",peer="127.0.0.3:80"} 0
  # HELP current_time current time in seconds
  # TYPE current_time counter
  current_time 536431200.095