API: fixed transition of a single peer from "unavailable" to "up".
If an upstream group contains only one peer, that peer is technically never
considered dead and is always selected.  So for this case there's a short path
in the peer selection code that avoids various checks and initializations.
But for the sake of accounting and display in our statistics API, we consider
sequential failures as the "unavailable" peer state.

Previously, the code path for a single peer didn't update the "peer->checked"
value, so transitioning to the "up" state was impossible because the
"peer->accessed < peer->checked" condition was always false.